requirements.txt
faster-whisper
sounddevice
numpy
groq
mcp[cli]
pyautogui
mss
pygetwindow
Pillow
scikit-learn
opencv-python
python-dotenv
